requirements.txt
# Core
streamlit>=1.28.0
python-pptx>=0.6.21
pillow>=10.1.0

# AI/ML
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.3
sentence-transformers>=2.2.2

# RAG и базы данных
chromadb>=0.4.18
langchain>=0.1.0
langchain-community>=0.1.0

# Fine-tuning
peft>=0.7.1
datasets>=2.15.0

# Документы
PyPDF2>=3.0.0
python-docx>=1.1.0

# Визуализация
matplotlib>=3.8.0
plotly>=5.18.0

# Оптимизация
xformers>=0.0.23
scikit-learn>=1.3.0